requirements.txt
accelerate==1.10.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.3.0
curl_cffi==0.11.1
deep-translator==1.11.4
diskcache==5.6.3
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.117.1
fastapi-cache2==0.2.2
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
googletrans==4.0.0rc1
greenlet==3.2.4
h11==0.16.0
h2==3.2.0
hf-xet==1.1.7
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.63
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.44
llama_cpp_python==0.3.15
load-dotenv==0.1.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multitasking==0.0.11
narwhals==2.5.0
neo4j==5.28.1
networkx==3.3
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.18.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
selenium==4.33.0
sentence-transformers==5.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.48.0
streamlit==1.50.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.5.1+cu121
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.37.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yfinance==0.2.66
zstandard==0.23.0
